+++ apps/cms/src/app/cms.module.ts
@@ -1,7 +1,61 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Route } from '@angular/router';
-import { createAdminRoutes } from '@blockframes/utils/routes/create-routes';
+import { AuthGuard } from '@blockframes/auth/guard/auth.guard';
+import { OrganizationGuard } from '@blockframes/organization/guard/organization.guard';
+import { PermissionsGuard } from '@blockframes/permissions/guard/permissions.guard';
+import { MaintenanceGuard } from '@blockframes/ui/maintenance/maintenance.guard';
+import { RouteOptions } from '@blockframes/utils/routes/create-routes';
 import { IdlePreload, IdlePreloadModule } from 'angular-idle-preload';
+
+export function createAdminRoutes({ appsRoutes }: RouteOptions) {
+  return [
+    {
+      path: 'maintenance',
+      canActivate: [MaintenanceGuard],
+      loadChildren: () => import('@blockframes/ui/maintenance/maintenance.module').then(m => m.MaintenanceModule)
+    },
+    {
+      path: '',
+      canActivate: [MaintenanceGuard],
+      children: [
+        { path: '', redirectTo: 'auth', pathMatch: 'full' },
+        {
+          path: 'auth',
+          loadChildren: () => import('@blockframes/auth/auth.module').then(m => m.AuthModule)
+        },
+        {
+          path: 'c',
+          canActivate: [AuthGuard],
+          canDeactivate: [AuthGuard],
+          children: [
+            {
+              path: '',
+              redirectTo: 'o',
+              pathMatch: 'full'
+            },
+            {
+              // The redirection route when user has no organization
+              path: 'organization',
+              loadChildren: () => import('@blockframes/organization/no-organization.module').then(m => m.NoOrganizationModule)
+            },
+            {
+              path: 'o',
+              canActivate: [PermissionsGuard, OrganizationGuard],
+              children: appsRoutes
+            }
+          ]
+        },
+        {
+          path: 'not-found',
+          loadChildren: () => import('@blockframes/ui/error/error-not-found.module').then(m => m.ErrorNotFoundModule)
+        },
+        {
+          path: '**',
+          loadChildren: () => import('@blockframes/ui/error/error-not-found.module').then(m => m.ErrorNotFoundModule)
+        }
+      ]
+    }];
+}
 
 const routes: Route[] = createAdminRoutes({
   appsRoutes: [{
