+++ libs/testing/cypress/browser/src/lib/helpers.ts
@@ -1,4 +1,4 @@
-import { PublicUser, PermissionsDocument, Organization } from '@blockframes/model';
+import { PublicUser, PermissionsDocument, Organization, User } from '@blockframes/model';
 import { USER_FIXTURES_PASSWORD } from '@blockframes/devops';
 import { serverId } from '@blockframes/utils/constants';
 import faker from '@faker-js/faker';
@@ -44,7 +44,11 @@ interface E2EUser {
   orgId?: string;
 }
 
-export const e2eUser = (data: E2EUser): PublicUser => {
+export const e2eUser = (data: E2EUser): User => {
+  const legalTerms = {
+    date: new Date(),
+    ip: '111.111.111.111'
+  };
   const { uid, firstName, lastName, email, orgId } = data;
   return {
     uid,
@@ -53,12 +57,24 @@ export const e2eUser = (data: E2EUser): PublicUser => {
     email,
     orgId,
     hideEmail: false,
+    privacyPolicy: legalTerms,
+    termsAndConditions: {
+      festival: legalTerms,
+      catalog: legalTerms
+    },
+    phoneNumber: '',
+    position: '',
+    financing: {
+      rank: ''
+    },
+    avatar: null,
     _meta: {
       emailVerified: true,
       createdFrom: 'festival',
       createdBy: 'anonymous',
       createdAt: now,
     },
+    
   };
 };
