+++ libs/utils/src/lib/routes/create-routes.ts
@@ -8,7 +8,7 @@ import { OrganizationGuard } from '@blockframes/organization/guard/organization.
 import { MaintenanceGuard } from '@blockframes/ui/maintenance';
 import { RequestAccessGuard } from '@blockframes/organization/guard/request-access.guard';
 
-interface RouteOptions {
+export interface RouteOptions {
   /** The routes of the apps */
   appsRoutes: Routes,
   /** The route to the landing page if any */
@@ -97,58 +97,5 @@ export function createRoutes({ appsRoutes, landing, events }: RouteOptions) {
       ]
     }]
 }
-
-// Used for CMS
-// Strip out the notifications / invitations
-export function createAdminRoutes({ appsRoutes }: RouteOptions) {
-  return [
-    {
-      path: 'maintenance',
-      canActivate: [MaintenanceGuard],
-      loadChildren: () => import('@blockframes/ui/maintenance/maintenance.module').then(m => m.MaintenanceModule)
-    },
-    {
-      path: '',
-      canActivate: [MaintenanceGuard],
-      children: [
-        { path: '', redirectTo: 'auth', pathMatch: 'full' },
-        {
-          path: 'auth',
-          loadChildren: () => import('@blockframes/auth/auth.module').then(m => m.AuthModule)
-        },
-        {
-          path: 'c',
-          canActivate: [AuthGuard],
-          canDeactivate: [AuthGuard],
-          children: [
-            {
-              path: '',
-              redirectTo: 'o',
-              pathMatch: 'full'
-            },
-            {
-              // The redirection route when user has no organization
-              path: 'organization',
-              loadChildren: () => import('@blockframes/organization/no-organization.module').then(m => m.NoOrganizationModule)
-            },
-            {
-              path: 'o',
-              canActivate: [PermissionsGuard, OrganizationGuard],
-              children: appsRoutes
-            }
-          ]
-        },
-        {
-          path: 'not-found',
-          loadChildren: () => import('@blockframes/ui/error/error-not-found.module').then(m => m.ErrorNotFoundModule)
-        },
-        {
-          path: '**',
-          loadChildren: () => import('@blockframes/ui/error/error-not-found.module').then(m => m.ErrorNotFoundModule)
-        }
-      ]
-    }];
-}
-
 
 export const appsRoute = '/c/o';
